refactor nmap nse page to dynamic client module
